AEGIS - WIP on HEDP rounds, notes in description
As it is right now, it is perfectly function, HOWEVER. It is basically a copy of the HE grenade due to a lack of penetrating power (low velocity), I tried altering the caliber to fix but at the moment it has very little actual penetrating power. Other than that, it isn't broken or buggy from what I've seen.